requirements.txt
pandas
torch
torchvision
Pillow
tqdm
transformers
sentence-transformers
safetensors
ultralytics
pymupdf
accelerate